+++ packages/moonstone/ExpandableList/ExpandableList.js
@@ -97,6 +97,23 @@ const ExpandableListBase = kind({
 		title: PropTypes.string.isRequired,
 
 		/**
+		 * Sets the hint string read when focusing the ExpandableList.
+		 *
+		 * @type {String}
+		 * @memberof moonstone/ExpandableList.prototype
+		 * @public
+		 */
+		'aria-label': PropTypes.string,
+
+		/**
+		 * Sets the hint string read when focusing the each item.
+		 *
+		 * @type {Array}
+		 * @public
+		 */
+		childrenAriaLabels: PropTypes.array,
+
+		/**
 		 * When `true` and `select` is not `'multiple'`, the expandable will be closed when an item
 		 * is selected.
 		 *
@@ -246,6 +263,7 @@ const ExpandableListBase = kind({
 	},
 
 	defaultProps: {
+		childrenAriaLabels: [],
 		select: 'radio',
 		spotlightDisabled: false
 	},
@@ -270,6 +288,34 @@ const ExpandableListBase = kind({
 
 	computed: {
 		'aria-multiselectable': ({select}) => select === 'multiple',
+
+		children: ({children, childrenAriaLabels}) => {
+			if (childrenAriaLabels && childrenAriaLabels.length === 0) {
+				return children;
+			}
+			const childrenArray = [];
+			children.map((element, i) => {
+				childrenArray.push({'aria-label': childrenAriaLabels[i], children: element, key: i});
+			});
+
+			return childrenArray;
+		},
+
+		computedAriaLabel: ({'aria-label': ariaLabel, childrenAriaLabels, label, select, selected, title}) => {
+			if (ariaLabel != null || (childrenAriaLabels && childrenAriaLabels.length === 0)) {
+				return null;
+			}
+
+			if (label) {
+				return `${title} ${label}`;
+			} else if (selected || selected === 0) {
+				const isArray = Array.isArray(selected);
+				if (select === 'multiple' && isArray) {
+					const selectedChildrenAriaLabel = selected.map(i => childrenAriaLabels[i]).filter(str => !!str).join(', ');
+					return `${title} ${selectedChildrenAriaLabel}`;
+				}
+			}
+		},
 
 		disabled: ({children, disabled}) => {
 			return disabled || !children || children.length === 0;
@@ -318,7 +364,9 @@ const ExpandableListBase = kind({
 	},
 
 	render: ({
+		'aria-label': ariaLabel,
 		children,
+		computedAriaLabel,
 		itemProps,
 		ListItem,
 		noAutoClose,
@@ -328,11 +376,13 @@ const ExpandableListBase = kind({
 		selected,
 		...rest
 	}) => {
+		delete rest.childrenAriaLabels;
 		delete rest.closeOnSelect;
 
 		return (
 			<ExpandableItemBase
 				{...rest}
+				aria-label={ariaLabel || computedAriaLabel}
 				showLabel="auto"
 				autoClose={!noAutoClose}
 				lockBottom={!noLockBottom}